Move tests out of module
